+++ langchain/agent.py
@@ -16,7 +16,7 @@ from psycopg_pool import AsyncConnectionPool
 import logging
 from rich.logging import RichHandler
 from rich.traceback import install
-from db_url import compose_postgres_url
+from helpers.db_url import compose_postgres_url
 from graph.context_loader import make_context_loader_node
 from graph.freeform import build_freeform_subgraph
 from graph.router import make_top_router_node
